+++ src/DataStore.js
@@ -379,10 +379,6 @@ function DataStore (opts) {
 const props = {
   constructor: DataStore,
 
-  _callSuper (method, ...args) {
-    return this.constructor.__super__.prototype[method].apply(this, args)
-  },
-
   /**
    * Internal method used to handle Mapper responses.
    *
@@ -857,7 +853,7 @@ const props = {
    */
   create (name, record, opts) {
     opts || (opts = {})
-    return this._callSuper('create', name, record, opts)
+    return Container.prototype.create.call(this, name, record, opts)
       .then((result) => this._end(name, result, opts))
   },
 
@@ -954,7 +950,7 @@ const props = {
    */
   createMany (name, records, opts) {
     opts || (opts = {})
-    return this._callSuper('createMany', name, records, opts)
+    return Container.prototype.createMany.call(this, name, records, opts)
       .then((result) => this._end(name, result, opts))
   },
 
@@ -1423,7 +1419,7 @@ const props = {
    */
   destroy (name, id, opts) {
     opts || (opts = {})
-    return this._callSuper('destroy', name, id, opts).then((result) => {
+    return Container.prototype.destroy.call(this, name, id, opts).then((result) => {
       const record = this.getCollection(name).remove(id, opts)
 
       if (record && this.unlinkOnDestroy) {
@@ -1536,7 +1532,7 @@ const props = {
    */
   destroyAll (name, query, opts) {
     opts || (opts = {})
-    return this._callSuper('destroyAll', name, query, opts).then((result) => {
+    return Container.prototype.destroyAll.call(this, name, query, opts).then((result) => {
       const records = this.getCollection(name).removeAll(query, opts)
 
       if (records && records.length && this.unlinkOnDestroy) {
@@ -1667,7 +1663,7 @@ const props = {
     let promise
 
     if (opts.force || !item) {
-      promise = this._pendingQueries[name][id] = this._callSuper('find', name, id, opts).then((result) => {
+      promise = this._pendingQueries[name][id] = Container.prototype.find.call(this, name, id, opts).then((result) => {
         delete this._pendingQueries[name][id]
         result = this._end(name, result, opts)
         this.cacheFind(name, result, id, opts)
@@ -1780,7 +1776,7 @@ const props = {
     let promise
 
     if (opts.force || !items) {
-      promise = this._pendingQueries[name][hash] = this._callSuper('findAll', name, query, opts).then((result) => {
+      promise = this._pendingQueries[name][hash] = Container.prototype.findAll.call(this, name, query, opts).then((result) => {
         delete this._pendingQueries[name][hash]
         result = this._end(name, result, opts)
         this.cacheFindAll(name, result, hash, opts)
@@ -2065,7 +2061,7 @@ const props = {
    */
   update (name, id, record, opts) {
     opts || (opts = {})
-    return this._callSuper('update', name, id, record, opts)
+    return Container.prototype.update.call(this, name, id, record, opts)
       .then((result) => this._end(name, result, opts))
   },
 
@@ -2157,7 +2153,7 @@ const props = {
    */
   updateAll (name, props, query, opts) {
     opts || (opts = {})
-    return this._callSuper('updateAll', name, query, props, opts)
+    return Container.prototype.updateAll.call(this, name, query, props, opts)
       .then((result) => this._end(name, result, opts))
   },
 
@@ -2249,7 +2245,7 @@ const props = {
    */
   updateMany (name, records, opts) {
     opts || (opts = {})
-    return this._callSuper('updateMany', name, records, opts)
+    return Container.prototype.updateMany.call(this, name, records, opts)
       .then((result) => this._end(name, result, opts))
   }
 }